feat(lamp): accept uint32 fade durations for long drift fades
Widens the receive→fade chain for OVERRIDE_COLORS to match the uint32
wire field added in Task 1. Apply and restore signatures, stored members,
the configurator's beginFade parameter and fadeDurationMs_ field, and the
PendingOverrideColors/PendingRestoreColors slots all move from uint16 to
uint32. Brightness-override path is unchanged.

Overflow analysis: ease() computes currentStep × 511 / duration × 511 / 511.
At the 3,600,000 ms ceiling: 3,600,000 × 511 = 1,839,600,000 < UINT32_MAX.
No uint64 intermediate needed; existing uint32 parameters in ease() are safe.

Stale "16-bit math is fine" comment in configurator::draw() removed — it
became false after widening.

Native test added: applies with 1,800,000 ms and asserts storage without
truncation (1,800,000 & 0xFFFF = 50,752 would be the wrong value); also
pins the LUT index arithmetic at the midpoint.

+++ software/lamp-os/src/behaviors/configurator.cpp
@@ -15,7 +15,7 @@ bool ConfiguratorBehavior::fadeActive(uint32_t nowMs) const {
 }
 
 void ConfiguratorBehavior::beginFade(const std::vector<Color>& targetColors,
-                                     uint16_t fadeDurationMs) {
+                                     uint32_t fadeDurationMs) {
   // Snapshot the fade-FROM endpoint from the configurator's PREVIOUS
   // intended output (the `colors` field, optionally lerped if we're
   // interrupting a fade in flight) — NOT from fb->buffer. fb->buffer
@@ -92,9 +92,6 @@ void ConfiguratorBehavior::draw() {
       fb->buffer[i] = colors[i];
     }
   } else {
-    // 16-bit math is fine — fadeDurationMs_ ≤ 65535 and elapsed is bounded
-    // by that. The fade() helper takes (start, end, steps, currentStep),
-    // matches the existing easeFrames-based call site.
     const uint32_t elapsed = now - fadeStartMs_;
     const uint32_t duration = fadeDurationMs_;
     for (int i = 0; i < fb->pixelCount; i++) {
