+++ software/lamp-os/src/components/network/show_receiver.hpp
@@ -59,7 +59,7 @@ struct PendingOverrideColors {
   uint8_t sourceMac[6];
   lamp_protocol::OverrideSurface surface;
   lamp_protocol::OverrideSource sourceKind;
-  uint16_t fadeDurationMs;
+  uint32_t fadeDurationMs;
   uint8_t numColors;
   Color colors[lamp_protocol::kMaxOverrideColorsPerFrame];
 };
@@ -68,7 +68,7 @@ struct PendingRestoreColors {
   uint8_t sourceMac[6];
   lamp_protocol::OverrideSurface surface;
   lamp_protocol::OverrideSource sourceKind;
-  uint16_t fadeDurationMs;
+  uint32_t fadeDurationMs;
 };
 
 struct PendingOverrideBrightness {
